+++ secbootctl/features/pmi.py
@@ -10,6 +10,7 @@ import shutil
 import sys
 import textwrap
 from pathlib import Path
+from typing import Optional
 
 from secbootctl.core import AppController, BaseSubcmdCreator, AppError
 from secbootctl.env import Env
@@ -17,59 +18,63 @@ from secbootctl.helpers.cli import CliPrintHelper
 
 
 class PmiController(AppController):
-    HOOK_TARGET_PATHS: dict = {
-        'pacman': Path('/etc/pacman.d/hooks')
-    }
+    PACMAN_HOOK_PATH: Path = Path('/etc/pacman.d/hooks')
 
     def install(self):
-        """Copies the hook file(s) for the configured package manager into the corresponding hook directory."""
+        """Copies the hook files for the configured package manager into the corresponding hook directories."""
         pm_name: str = self._config.package_manager_name
-        hook_path: Path = Env.APP_HOOK_PATH / pm_name
 
         self._check_package_manager(pm_name)
 
         self._print_status(f'installing hook files for package manager: {pm_name}')
 
-        hook_target_path: Path = self.HOOK_TARGET_PATHS.get(pm_name)
-
-        for hook_file_path in glob.glob(str(hook_path / '*.*')):
-            hook_file_path: Path = Path(hook_file_path)
-            shutil.copy(hook_file_path, hook_target_path)
-            copied_hook_file_path: Path = hook_target_path / hook_file_path.name
-            shutil.chown(copied_hook_file_path, 'root', 'root')
-            os.chmod(copied_hook_file_path, 0o700)
+        getattr(self, '_' + pm_name + '_install')()
 
         self._print_status(f'installed hook files for package manager: {pm_name}', CliPrintHelper.Status.SUCCESS)
 
     def remove(self):
-        """Removes the hook file(s) for the configured package manager in the corresponding hook directory."""
+        """Removes the hook files for the configured package manager in the corresponding hook directories."""
         pm_name: str = self._config.package_manager_name
-        hook_path: Path = Env.APP_HOOK_PATH / pm_name
 
         self._check_package_manager(pm_name)
 
         self._print_status(f'removing hook files for package manager: {pm_name}')
 
-        hook_target_path: Path = self.HOOK_TARGET_PATHS.get(pm_name)
-
-        for hook_file_path in glob.glob(str(hook_path / '*.*')):
-            hook_file_path: Path = Path(hook_file_path)
-            target_hook_file_path: Path = hook_target_path / hook_file_path.name
-            target_hook_file_path.unlink(missing_ok=True)
+        getattr(self, '_' + pm_name + '_remove')()
 
         self._print_status(f'removed hook files for package manager: {pm_name}', CliPrintHelper.Status.SUCCESS)
 
-    def hook_callback(self, mode: str):
+    def hook_callback(self, mode: str, kernel_name: Optional[str] = None):
         """Callback invoked by package manager hook(s)."""
         pm_name: str = self._config.package_manager_name
 
         self._check_package_manager(pm_name)
 
-        getattr(self, '_' + pm_name + '_' + mode + '_callback')()
+        # @todo there is propably a prettier solution...
+        if pm_name == 'pacman':
+            getattr(self, '_' + pm_name + '_' + mode + '_callback')()
+        else:
+            getattr(self, '_' + pm_name + '_' + mode + '_callback')(kernel_name)
 
     def _check_package_manager(self, pm_name: str) -> None:
         if pm_name not in Env.SUPPORTED_PACKAGE_MANAGERS:
             raise AppError(f'configured package manager "{pm_name}" is not supported')
+
+    def _pacman_install(self):
+        pm_name: str = self._config.package_manager_name
+        hook_path: Path = Env.APP_HOOK_PATH / pm_name
+
+        for hook_file_path in glob.glob(str(hook_path / '*.*')):
+            self._copy_hook_file(Path(hook_file_path), self.PACMAN_HOOK_PATH)
+
+    def _pacman_remove(self):
+        pm_name: str = self._config.package_manager_name
+        hook_path: Path = Env.APP_HOOK_PATH / pm_name
+
+        for hook_file_path in glob.glob(str(hook_path / '*.*')):
+            hook_file_path: Path = Path(hook_file_path)
+            target_hook_file_path: Path = self.PACMAN_HOOK_PATH / hook_file_path.name
+            target_hook_file_path.unlink(missing_ok=True)
 
     def _pacman_update_callback(self):
         boot_path: Path = self._config.boot_path
@@ -96,6 +101,35 @@ class PmiController(AppController):
 
             self._forward('kernel', 'remove', {'kernel_name': kernel_name})
 
+    def _apt_install(self):
+        pm_name: str = self._config.package_manager_name
+        hook_path: Path = Env.APP_HOOK_PATH / pm_name
+
+        self._copy_hook_file(hook_path / 'initramfs' / 'yy-secbootctl-update', Path('/etc/initramfs/post-update.d'))
+        self._copy_hook_file(hook_path / 'kernel' / 'yy-secbootctl-update', Path('/etc/kernel/postinst.d'))
+        self._copy_hook_file(hook_path / 'kernel' / 'yy-secbootctl-remove', Path('/etc/kernel/postrm.d'))
+
+    def _apt_remove(self):
+        Path('/etc/initramfs/post-update.d/yy-secbootctl-update').unlink(missing_ok=True)
+        Path('/etc/kernel/postinst.d/yy-secbootctl-update').unlink(missing_ok=True)
+        Path('/etc/kernel/postrm.d/yy-secbootctl-remove').unlink(missing_ok=True)
+
+    def _apt_update_callback(self, kernel_name: str):
+        # @todo what to do with systemd-boot updates?
+        self._forward('kernel', 'install', {'kernel_name': kernel_name})
+
+    def _apt_remove_callback(self, kernel_name: str):
+        self._forward('kernel', 'remove', {'kernel_name': kernel_name})
+
+    def _copy_hook_file(self, hook_file_path: Path, target_hook_path: Path):
+        if not target_hook_path.is_dir():
+            os.makedirs(target_hook_path, 0o755, True)
+
+        shutil.copy(hook_file_path, target_hook_path)
+        copied_hook_file_path: Path = target_hook_path / hook_file_path.name
+        shutil.chown(copied_hook_file_path, 'root', 'root')
+        os.chmod(copied_hook_file_path, 0o700)
+
 
 class PmiSubcmdCreator(BaseSubcmdCreator):
     def create(self, cli_subparsers):
@@ -114,3 +148,4 @@ class PmiSubcmdCreator(BaseSubcmdCreator):
             Callback for package manager hook. Gets invoked by the package manager hook itself.
         '''))
         pmc_cli_subparser.add_argument('mode', help='e.g. "update", "remove", etc.')
+        pmc_cli_subparser.add_argument('kernel_name', nargs='?', help='e.g. "5.4.0-91-generic", etc.')